# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Assistant API",
    "description": "AI assistant API — A2A protocol (Agent-to-Agent), chat, and workflow management.\n\n**Authentication**: protected endpoints require `Authorization: Bearer <token>`.\nThe token is set via `--auth-token` / `ASSISTANT_WEB_TOKEN` on the server.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.135"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local assistant server"
    }
  ],
  "paths": {
...
    "/api/orgs/{org_id}/catalog/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "`GET /api/orgs/{org_id}/catalog/templates` — list available persona templates.",
        "operationId": "list_templates",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_token": []
          },
          {
            "oauth2": []
          }
        ]
      }
    },
...
    "/api/orgs/{org_id}/spaces/{space_id}/personas/from-template": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "`POST /api/orgs/{org_id}/spaces/{space_id}/personas/from-template` — create a persona from a template.",
        "operationId": "create_from_template",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "space_id",
            "in": "path",
            "description": "Space ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFromTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Persona created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaFromTemplateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Template not found"
          }
        },
        "security": [
          {
            "bearer_token": []
          },
          {
            "oauth2": [
              "spaces:write"
            ]
          }
        ]
      }
    },
...
    "/api/users/me/onboarding-status": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "`GET /api/users/me/onboarding-status` — check if user has created at least one persona.",
        "operationId": "onboarding_status",
        "responses": {
          "200": {
            "description": "Onboarding status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          },
          {
            "oauth2": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateFromTemplateRequest": {
        "type": "object",
        "required": [
          "template_id"
        ],
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name for the new persona (optional — defaults to template name)."
          },
          "template_id": {
            "type": "string",
            "description": "Catalog item ID of the template to instantiate."
          }
        }
      },
...
      "OnboardingStatusResponse": {
        "type": "object",
        "required": [
          "has_persona"
        ],
        "properties": {
          "has_persona": {
            "type": "boolean",
            "description": "Whether the user has created at least one persona."
          }
        }
      },
...
      "PersonaFromTemplateResponse": {
        "type": "object",
        "required": [
          "persona_id",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "persona_id": {
            "type": "string"
          }
        }
      },
...
      "TemplateResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token. Pass the token issued by the server via `Authorization: Bearer <token>`. Configure it with `--auth-token` / `ASSISTANT_WEB_TOKEN`."
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "refreshUrl": "/oauth/token",
            "scopes": {
              "api_keys:read": "List API keys",
              "api_keys:write": "Create and revoke API keys",
              "conversations:read": "Read conversations",
              "conversations:write": "Create and send messages",
              "org:manage": "Full organization management",
              "personas:read": "Read personas",
              "personas:write": "Create and update personas",
              "skills:read": "Read skills",
              "skills:write": "Create and update skills",
              "spaces:read": "Read spaces",
              "spaces:write": "Create and update spaces",
              "users:read": "Read users",
              "users:write": "Create and update users"
            }
          },
          "clientCredentials": {
            "tokenUrl": "/oauth/token",
            "scopes": {
              "api_keys:read": "List API keys",
              "api_keys:write": "Create and revoke API keys",
              "conversations:read": "Read conversations",
              "conversations:write": "Create and send messages",
              "org:manage": "Full organization management",
              "personas:read": "Read personas",
              "personas:write": "Create and update personas",
              "skills:read": "Read skills",
              "skills:write": "Create and update skills",
...
  },
...
}